+++ src/editor/core/command/CommandAdapt.ts
@@ -64,6 +64,7 @@ import {
   IElementPosition,
   IElementStyle,
   IGetElementByIdOption,
+  IInsertElementListOption,
   IUpdateElementByIdOption
 } from '../../interface/Element'
 import {
@@ -1375,6 +1376,10 @@ export class CommandAdapt {
     return this.draw.getValue(options)
   }
 
+  public getValueAsync(options?: IGetValueOption): Promise<IEditorResult> {
+    return this.draw.getWorkerManager().getValue(options)
+  }
+
   public getAreaValue(
     options?: IGetAreaValueOption
   ): IGetAreaValueResult | null {
@@ -1441,10 +1446,14 @@ export class CommandAdapt {
     const endElement = pickElementAttr(elementList[endIndex], {
       extraPickAttrs: ['id']
     })
-    // 页码信息
+    // 页码信息、行信息
     const positionList = this.position.getPositionList()
-    const startPageNo = positionList[startIndex].pageNo
-    const endPageNo = positionList[endIndex].pageNo
+    const startPosition = positionList[startIndex]
+    const endPosition = positionList[endIndex]
+    const startPageNo = startPosition.pageNo
+    const endPageNo = endPosition.pageNo
+    const startRowNo = startPosition.rowIndex
+    const endRowNo = endPosition.rowIndex
     // 坐标信息（相对编辑器书写区）
     const rangeRects: RangeRect[] = []
     const height = this.draw.getOriginalHeight()
@@ -1531,6 +1540,8 @@ export class CommandAdapt {
       endElement,
       startPageNo,
       endPageNo,
+      startRowNo,
+      endRowNo,
       rangeRects,
       zone,
       isTable,
@@ -1653,10 +1664,18 @@ export class CommandAdapt {
     })
   }
 
-  public insertElementList(payload: IElement[]) {
+  public insertElementList(
+    payload: IElement[],
+    options: IInsertElementListOption = {}
+  ) {
     if (!payload.length) return
     const isDisabled = this.draw.isReadonly() || this.draw.isDisabled()
     if (isDisabled) return
+    const { isReplace = true } = options
+    // 如果配置不替换时，需收缩选区至末尾
+    if (!isReplace) {
+      this.range.shrinkRange()
+    }
     const cloneElementList = deepClone(payload)
     // 格式化上下文信息
     const { startIndex } = this.range.getRange()
@@ -2026,15 +2045,27 @@ export class CommandAdapt {
   }
 
   public setControlValue(payload: ISetControlValueOption) {
-    this.draw.getControl().setValueById(payload)
+    this.draw.getControl().setValueListById([payload])
+  }
+
+  public setControlValueList(payload: ISetControlValueOption[]) {
+    this.draw.getControl().setValueListById(payload)
   }
 
   public setControlExtension(payload: ISetControlExtensionOption) {
-    this.draw.getControl().setExtensionById(payload)
+    this.draw.getControl().setExtensionListById([payload])
+  }
+
+  public setControlExtensionList(payload: ISetControlExtensionOption[]) {
+    this.draw.getControl().setExtensionListById(payload)
   }
 
   public setControlProperties(payload: ISetControlProperties) {
-    this.draw.getControl().setPropertiesById(payload)
+    this.draw.getControl().setPropertiesListById([payload])
+  }
+
+  public setControlPropertiesList(payload: ISetControlProperties[]) {
+    this.draw.getControl().setPropertiesListById(payload)
   }
 
   public setControlHighlight(payload: ISetControlHighlightOption) {