+++ src/editor/dataset/constant/Cursor.ts
@@ -6,5 +6,6 @@ export const defaultCursorOption: Readonly<Required<ICursorOption>> = {
   width: 1,
   color: '#000000',
   dragWidth: 2,
-  dragColor: '#0000FF'
+  dragColor: '#0000FF',
+  dragFloatImageDisabled: false
 }