+++ src/editor/interface/EventBus.ts
@@ -2,6 +2,7 @@ import {
   IContentChange,
   IControlChange,
   IControlContentChange,
+  IImageSizeChange,
   IIntersectionPageNoChange,
   IMouseEventChange,
   IPageModeChange,
@@ -33,4 +34,5 @@ export interface EventBusMap {
   mouseup: IMouseEventChange
   click: IMouseEventChange
   positionContextChange: IPositionContextChange
+  imageSizeChange: IImageSizeChange
 }